eliminate partial def
Until well founded recursion is supported, we have to prove them
manually here using the `partial/implementedBy` pattern. I've tried to
keep the definitions small so that the `implementedBy` match is
verifiable by inspection (we won't be able to prove correspondence).
The pure defs that remain are now amenable to proving properties.